% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{all_karyotype}
\alias{all_karyotype}
\title{This is a test data for this package
test data describtion}
\format{
An object of class \code{data.frame} with 88 rows and 6 columns.
}
\usage{
all_karyotype
}
\description{
This is a test data for this package
test data describtion
}
\author{
JunZhang
}
\keyword{datasets}
