% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGraphFuncs.R
\name{annotation_custom2}
\alias{annotation_custom2}
\title{Add a custom annotation layer to a ggplot}
\usage{
annotation_custom2(
  grob,
  xmin = -Inf,
  xmax = Inf,
  ymin = -Inf,
  ymax = Inf,
  data
)
}
\arguments{
\item{grob}{A grob object that will be used as the custom annotation.}

\item{xmin}{The minimum x value where the custom annotation will be placed (default is -Inf).}

\item{xmax}{The maximum x value where the custom annotation will be placed (default is Inf).}

\item{ymin}{The minimum y value where the custom annotation will be placed (default is -Inf).}

\item{ymax}{The maximum y value where the custom annotation will be placed (default is Inf).}

\item{data}{The data that the custom annotation layer will be applied to.}
}
\value{
A ggplot layer with the custom annotation.
}
\description{
This function adds a custom annotation layer to a ggplot using the provided grob object.
The annotation can be placed within specified x and y limits.
}
