% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcuCentromereData}
\alias{calcuCentromereData}
\title{Calculate the centromere data for karyotype}
\usage{
calcuCentromereData(karyotype = NULL)
}
\arguments{
\item{karyotype}{A data frame with four columns：chr, size, ce_start, ce_end.
The columns represent the chromosome number, size of the chromosome, start
position of the centromere, and end position of the centromere, respectively.}
}
\value{
a data frame which contains three columns: chr, x, and y. x and y
represent the coordinate for visualization. The data frame is used for centromere
visualization in karyotype plot.
}
\description{
Generate data for centromere visualization in karyotype plot.
}
