% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcuIntervals}
\alias{calcuIntervals}
\title{Calculate the segmented interval for chromosome}
\usage{
calcuIntervals(chr, length, step = 2 * 10^6)
}
\arguments{
\item{chr}{A character which represents the chromosome of segmented intervals.}

\item{length}{A numerical value which represents the length of the chromosome.}

\item{step}{A numerical value which represents the length of each segmentation,
default is 1,000,000bp.}
}
\value{
a dataframe which includes three columns representing the chromosome
number, starting coordinate, and ending coordinate of each segmented interval.
}
\description{
Generate segmented intervals for chromosome based on the total length and
specified segmentation size.
}
