% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchIdeoData.R
\name{catchIdeoData}
\alias{catchIdeoData}
\title{Download and process cytoband ideogram data}
\usage{
catchIdeoData(genome = "hg19", chromosome = NULL)
}
\arguments{
\item{genome}{The genome assembly version (default is "hg19")}

\item{chromosome}{The chromosome number to filter by (default is NULL)}
}
\value{
A list containing the processed ideogram data
}
\description{
\itemize{
\item "gneg": 染色体中的负色区域。这些区域通常不含有基因，且在染色过程中呈现较浅的颜色.
\item "gpos25": 染色体中呈现较深颜色的区域，具有25\%的染色强度。这些区域通常含有基因.
\item "gpos50": 染色体中呈现较深颜色的区域，具有50\%的染色强度。这些区域通常含有基因.
\item "gpos75": 染色体中呈现较深颜色的区域，具有75\%的染色强度。这些区域通常含有基因.
\item "gpos100": 染色体中呈现最深颜色的区域，具有100\%的染色强度。这些区域通常含有基因.
\item "acen": 染色体的着丝点。这是染色体中的一个特殊区域，负责染色体在细胞分裂过程中的移动与排列.
\item "gvar": 染色体中的变异区域。这些区域在不同个体或种群之间具有较高的基因变异率.
\item "stalk": 染色体的柄区。这是染色体中的一个狭窄区域，连接着着丝点和其他染色质区域.
}
}
\details{
This function downloads cytoband ideogram data from the UCSC genome browser
and assigns colors to different regions based on their densities. It then prepares the
data for plotting by creating border and centromere connection data, as well as a plot
dataframe. The resulting data is returned as a list.
}
\examples{
\dontrun{ideogram_data <- catchIdeoData(genome = "hg19")}

}
\author{
JunZhang
}
