% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createPairPolygon}
\alias{createPairPolygon}
\title{Create a Pair of Polygons}
\usage{
createPairPolygon(
  start = NULL,
  end = NULL,
  open = 1,
  npoints = 1000,
  reverse = FALSE
)
}
\arguments{
\item{start}{A numeric value indicating the start point of the polygon.
Defaults to NULL.}

\item{end}{A numeric value indicating the end point of the polygon. Defaults
to NULL.}

\item{open}{A numeric value indicating the openness of the polygon. Defaults
to 1.}

\item{npoints}{An integer value indicating the number of points to be used to
construct the polygon. Defaults to 1000.}

\item{reverse}{A logical value indicating whether to reverse the order of the
polygons.}
}
\value{
A data.frame object containing the x and y coordinates of the two
polygons and the type of polygon.
}
\description{
This function creates a pair of polygons, each consisting of two line segments,
with a shared mid-point and a user-defined start and end point.
}
\examples{
createPairPolygon(start = 0, end = 10, open = 1, npoints = 1000, reverse = FALSE)

}
