% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createSegment}
\alias{createSegment}
\title{Create segment coordinates}
\usage{
createSegment(
  xPos = NULL,
  yPos = NULL,
  n_division = NULL,
  rel_len = NULL,
  abs_len = NULL
)
}
\arguments{
\item{xPos}{A vector of length 2 containing the x-coordinates of the starting
and ending positions for the segment.}

\item{yPos}{A vector of length 2 containing the y-coordinates of the starting
and ending positions for the segment.}

\item{n_division}{An integer value indicating the number of segments the line
should be divided into.}

\item{rel_len}{A decimal value indicating the relative length of each segment.}

\item{abs_len}{A absolute length for generating segments.}
}
\value{
A data frame with the x and y-coordinates of each segment.
}
\description{
This function takes starting and ending coordinates and creates segments in
equal divisions or given relative length. It returns a data frame with the x
and y coordinates of each segment.
}
\examples{
createSegment(xPos = c(0.5, 0.8), yPos = c(0.2, 0.5), n_division = 4)
createSegment(xPos = c(0.5, 0.8), yPos = c(0.2, 0.5), rel_len = 0.2)
createSegment(xPos = c(0.5, 0.8), yPos = c(0.2, 0.5), abs_len = 0.1)

}
