% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGraphFuncs.R
\name{createTrapezoid}
\alias{createTrapezoid}
\title{Create Trapezoid Shape}
\usage{
createTrapezoid(xPos = NULL, yPos = NULL, nDivisions = 2000, reverse = FALSE)
}
\arguments{
\item{xPos}{A numeric vector of length 4 representing x-coordinates of the
trapezoid in order.}

\item{yPos}{A numeric vector of length 4 representing y-coordinates of the
trapezoid in order.}

\item{nDivisions}{An integer value indicating the number of divisions to
create the trapezoid shape. Default is 2000.}

\item{reverse}{A boolean value indicating whether to create the trapezoid in
reverse order or not. Default is FALSE.}
}
\value{
A data frame with id, x and y coordinates.
}
\description{
This function creates a trapezoid shape with given coordinates.
}
