% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGraphFuncs.R
\name{createTriangle}
\alias{createTriangle}
\title{Create Triangle Coordinates}
\usage{
createTriangle(x = NULL, y = NULL, nDivisions = 1, vertice = 2)
}
\arguments{
\item{x}{numeric vector of length 3, containing the x coordinates of the
triangle vertices.}

\item{y}{numeric vector of length 3, containing the y coordinates of the
triangle vertices.}

\item{nDivisions}{integer specifying the number of divisions between each
vertex. Default is 1.}

\item{vertice}{integer specifying which vertex to use as the base. Must be 1,
2 or 3. Default is 2.}
}
\value{
A data frame with columns "x", "y" and "id". The "id" column specifies
which division the coordinate belongs to.
}
\description{
This function creates a data frame of x and y coordinates for a triangle
defined by three vertices.
The user can specify the number of divisions and which vertex to use as the base.
}
\examples{
\dontrun{
createTriangle(x = c(0, 5, 3), y = c(0, 0, 4))
createTriangle(x = c(0, 5, 3), y = c(0, 0, 4), nDivisions = 3, vertice = 1)
}

}
