% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawChromosome.R
\name{drawChromosome}
\alias{drawChromosome}
\title{Draw Chromosome Karyotype Plot}
\usage{
drawChromosome(
  ideogram_obj = NULL,
  karyotype_file = NULL,
  centromereTri_file = NULL,
  density_file = NULL,
  chromosomes = NULL,
  orders = NULL,
  density_col = c("grey90", "black"),
  centromere_col = "#F00000",
  zoom_region = NULL,
  zoom_pos = "top",
  zoom_col = c("grey95", "grey60"),
  zoom_relheight = 1,
  facet_params = list(),
  add_regionLen = TRUE,
  ...
)
}
\arguments{
\item{ideogram_obj}{The output for "catchIdeoData" function.}

\item{karyotype_file}{Path to the karyotype file. The file is expected to be
in tab-delimited format with columns chr, start, end, and name representing
chromosome name, start and end positions, and chromosome band name respectively.}

\item{centromereTri_file}{Path to the centromere triangle file. The file is
expected to be in tab-delimited format with columns chr, start, and end
representing chromosome name, start and end positions of the centromere
triangle respectively.}

\item{density_file}{Path to the density file. The file is expected to be in
tab-delimited format with columns chr, pos, and density representing chromosome
name, position and density value.}

\item{chromosomes}{A character vector indicating the chromosomes to be
displayed in the plot.}

\item{orders}{A named integer vector indicating the order of chromosomes to be
displayed in the plot.}

\item{density_col}{A character vector of length 2 indicating the colors to be
used for the density plot.}

\item{centromere_col}{A character value indicating the color to be used for
the centromere triangles.}

\item{zoom_region}{A numeric vector of length 2 indicating the region to be
zoomed in. Default is NULL (no zoom).}

\item{zoom_pos}{A character value indicating the location of the zoomed region.
Possible values are "top" or "bottom".}

\item{zoom_col}{A character vector of length 2 indicating the colors to be
used for the zoomed region.}

\item{zoom_relheight}{A numeric value indicating the relative height of the
zoomed region with respect to the main plot.}

\item{facet_params}{List of parameters for facet_wrap function.}

\item{add_regionLen}{Whether add region length information.}

\item{...}{Additional parameters to be passed to the theme function.}
}
\description{
Draw Chromosome Karyotype Plot
}
\author{
JunZhang

This function creates a chromosome karyotype plot from the given input files
and parameters.
}
