% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{filterID,GenomeGTF-method}
\alias{filterID,GenomeGTF-method}
\title{Filter genome annotation information by gene name, gene ID, or transcript ID}
\usage{
\S4method{filterID}{GenomeGTF}(object, geneName = NULL, geneId = NULL, transId = NULL)
}
\arguments{
\item{object}{A GenomeGTF object.}

\item{geneName}{A character vector of gene names to filter by. Default is NULL.}

\item{geneId}{A character vector of gene IDs to filter by. Default is NULL.}

\item{transId}{A character vector of transcript IDs to filter by. Default is NULL.}
}
\value{
A data.frame containing the filtered genome annotation information
}
\description{
This function filters the genome annotation information in a GenomeGTF object
based on the provided gene name, gene ID, or transcript ID.
}
