% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{filterID}
\alias{filterID}
\title{filterID method for GenomeGTF objects}
\usage{
filterID(object, ...)
}
\arguments{
\item{object}{A GenomeGTF object}

\item{...}{Other arguments.}
}
\description{
filterID method for GenomeGTF objects
}
