% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{filterRepTrans,GenomeGTF-method}
\alias{filterRepTrans,GenomeGTF-method}
\title{Filter redundant transcripts and retain the longest ones based on gene ID or
gene name}
\usage{
\S4method{filterRepTrans}{GenomeGTF}(
  object,
  geneName = NULL,
  geneId = NULL,
  selecType = c("lcds", "lt"),
  topN = 1,
  sep = "|"
)
}
\arguments{
\item{object}{A GenomeGTF object or a data.frame with GTF
format.}

\item{geneName}{A character vector of gene names. Default is NULL.}

\item{geneId}{A character vector of gene IDs. Default is NULL.}

\item{selecType}{A character vector specifying the selection criteria.
Must be either "lt" for longest transcript or "lcds" for longest CDS. Default
is "lcds".}

\item{topN}{An integer indicating the number of longest transcripts to retain.
Default is 1.
Set to 0 to retain all transcripts.}

\item{sep}{A character specifying the separator used to separate gene name and
transcript name.
Default is "|".}
}
\value{
A data.frame containing the longest transcripts for each gene based on
the selected criteria.
}
\description{
This function filters redundant transcripts based on gene ID or gene name,
and retains the longest ones according to the selected criteria.
}
