% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{filterRepTrans}
\alias{filterRepTrans}
\title{filterRepTrans method for GenomeGTF objects}
\usage{
filterRepTrans(object, ...)
}
\arguments{
\item{object}{A GenomeGTF object}

\item{...}{Other arguments.}
}
\description{
filterRepTrans method for GenomeGTF objects
}
