% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{GenomeGTF-class}
\alias{GenomeGTF-class}
\title{GenomeGTF}
\value{
return a GenomeGTF class.
}
\description{
GenomeGTF
}
\section{Slots}{

\describe{
\item{\code{gtf}}{GRanges. The GTF/GFF annotation file.}

\item{\code{genome}}{DNAStringSet. Genome sequence.}

\item{\code{representTrans}}{data.frame. longest transcript information.}

\item{\code{intron}}{data.frame. The intron data.}

\item{\code{ranges}}{GroupedIRanges. Null.}
}}

\author{
JunZhang
}
