% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getChromSize}
\alias{getChromSize}
\title{Get chromosome sizes from a genome object}
\usage{
getChromSize(object)
}
\arguments{
\item{object}{A Genome object}
}
\value{
A data.frame with columns for chromosome name, start and end positions.
}
\description{
This function takes a Genome object and returns a data.frame containing the
sizes of the chromosomes.
}
