% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getFeatureFromGenome,GenomeGTF-method}
\alias{getFeatureFromGenome,GenomeGTF-method}
\title{Extract genomic features from GenomeGTF object}
\usage{
\S4method{getFeatureFromGenome}{GenomeGTF}(
  object,
  nameData = NULL,
  sep = "|",
  geneName = NULL,
  geneId = NULL,
  transId = NULL,
  type = c("5UTR", "five_prime_utr", "3UTR", "three_prime_utr", "exon", "CDS", "intron"),
  geneSeq = FALSE,
  up.extend = 0,
  dn.extend = 0
)
}
\arguments{
\item{object}{A GenomeGTF object.}

\item{nameData}{A character vector or a data frame specifying the transcript
IDs to be extracted. If NULL, the function will use the geneName, geneId, and
transId arguments to filter the transcripts to be extracted. Default is NULL.}

\item{sep}{A character string specifying the separator used to concatenate
the gene name, gene ID, and transcript ID into a single name for the output
sequences. Default is "|".}

\item{geneName}{A character vector specifying the gene names to filter the
transcripts to be extracted. Default is NULL.}

\item{geneId}{A character vector specifying the gene IDs to filter the
transcripts to be extracted. Default is NULL.}

\item{transId}{A character vector specifying the transcript IDs to filter the
transcripts to be extracted. Default is NULL.}

\item{type}{A character string specifying the type of genomic features to be
extracted. Can be one of "5UTR", "five_prime_utr", "3UTR", "three_prime_utr",
"exon", "CDS", or "intron".}

\item{geneSeq}{A logical value specifying whether to extract the entire gene
sequence instead of the feature sequence. Default is FALSE.}

\item{up.extend}{The extend base for upstream of feature type. Default is 0.}

\item{dn.extend}{The extend base for downstream of feature type. Default is 0.}
}
\value{
A \code{DNAStringSet} object containing the genomic sequences of the
extracted features.
}
\description{
This function extracts genomic features, such as 5' UTR, 3' UTR, exons, CDS and
introns, from a GenomeGTF object and returns the corresponding genomic
sequences. The genomic sequences are obtained by fetching the genome sequences
using the coordinates of the features.
}
