% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getFeatureFromGenome}
\alias{getFeatureFromGenome}
\title{getFeatureFromGenome method for GenomeGTF objects}
\usage{
getFeatureFromGenome(object, ...)
}
\arguments{
\item{object}{A GenomeGTF object}

\item{...}{Other arguments.}
}
\description{
getFeatureFromGenome method for GenomeGTF objects
}
