% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGeneDensiy}
\alias{getGeneDensiy}
\title{Get gene density data for karyotype object}
\usage{
getGeneDensiy(karyotype = NULL, object = NULL, step = 2 * 10^6)
}
\arguments{
\item{karyotype}{A data frame with three columns: chr, start, and end. The
columns represent the chromosome number, start position and end position of
the chromosome, respectively.}

\item{object}{A character which contains the path of a gtf file.}

\item{step}{A numerical value which represents the length of each segmentation,
default is 1,000,000bp.}
}
\value{
a data frame which contains four columns: chr, start, end, and numGenes.
numGenes represents the number of genes in the corresponding segment.
}
\description{
Get gene density data for a list object of karyotype or a gtf file.
}
