% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getIntronInfo,GenomeGTF-method}
\alias{getIntronInfo,GenomeGTF-method}
\title{Get intron information from a GenomeGTF object}
\usage{
\S4method{getIntronInfo}{GenomeGTF}(object, geneName = NULL, geneId = NULL, transId = NULL)
}
\arguments{
\item{object}{A GenomeGTF object}

\item{geneName}{A character string specifying the gene name for which to
extract intron information. Default is NULL.}

\item{geneId}{A character string specifying the gene ID for which to extract
intron information. Default is NULL.}

\item{transId}{A character string specifying the transcript ID for which to
extract intron information. Default is NULL.}
}
\value{
A data.frame containing the start and end coordinates of each intron,
as well as other information such as the transcript ID and strand.
}
\description{
This function extracts intron information from a GenomeGTF object for a
given gene or transcript.
}
