% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getIntronInfo}
\alias{getIntronInfo}
\title{getIntronInfo method for GenomeGTF objects}
\usage{
getIntronInfo(object, ...)
}
\arguments{
\item{object}{A GenomeGTF object}

\item{...}{Other arguments.}
}
\description{
getIntronInfo method for GenomeGTF objects
}
