% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getLongName,GenomeGTF-method}
\alias{getLongName,GenomeGTF-method}
\title{Get long name of transcripts in GenomeGTF object}
\usage{
\S4method{getLongName}{GenomeGTF}(object, geneName = NULL, geneId = NULL, transId = NULL, sep = "|")
}
\arguments{
\item{object}{A GenomeGTF object.}

\item{geneName}{Character vector of gene names to filter by. Default is NULL.}

\item{geneId}{Character vector of gene IDs to filter by. Default is NULL.}

\item{transId}{Character vector of transcript IDs to filter by. Default is NULL.}

\item{sep}{Separator character to use in the long name. Default is "|".}
}
\value{
A data frame with columns gid, tid, and tname.
}
\description{
This function extracts information about transcripts in a GenomeGTF object
and returns a data frame with their long names.
}
