% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getNonRedundantLength,GenomeGTF-method}
\alias{getNonRedundantLength,GenomeGTF-method}
\title{Get the non-redundant length of transcripts for gene}
\usage{
\S4method{getNonRedundantLength}{GenomeGTF}(object, geneName = NULL, geneId = NULL, type = "exon")
}
\arguments{
\item{object}{A GenomeGTF object.}

\item{geneName}{A character string specifying the name of the gene for which
to retrieve the non-redundant length of transcripts. If NULL, the function will use geneId instead.}

\item{geneId}{A character string specifying the ID of the gene for which to
retrieve the non-redundant length of transcripts. If NULL, the function will use geneName instead.}

\item{type}{The feature type selected to calculate. Default "exon".}
}
\value{
A data frame for gene with non-redundant length.
}
\description{
This function returns the non-redundant length of transcripts for a specific
gene based on the information
}
\examples{
\dontrun{
# get the non-redundant length of transcripts for gene "ENSG00000139618"
getNonRedundantLength(GenomeGTFobject, geneId = "ENSG00000139618")
}

}
