% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getNonRedundantLength}
\alias{getNonRedundantLength}
\title{getNonRedundantLength method for GenomeGTF objects}
\usage{
getNonRedundantLength(object, ...)
}
\arguments{
\item{object}{A GenomeGTF object}

\item{...}{Other arguments.}
}
\description{
getNonRedundantLength method for GenomeGTF objects
}
