% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getPromoters,GenomeGTF-method}
\alias{getPromoters,GenomeGTF-method}
\title{Retrieve promoter regions for specified genes or transcripts.}
\usage{
\S4method{getPromoters}{GenomeGTF}(
  object,
  geneName = NULL,
  geneId = NULL,
  transId = NULL,
  up.extend = 2000,
  dn.extend = 0
)
}
\arguments{
\item{object}{A GenomeGTF object.}

\item{geneName}{A character vector of gene names to retrieve promoter regions for.}

\item{geneId}{A character vector of gene IDs to retrieve promoter regions for.}

\item{transId}{A character vector of transcript IDs to retrieve promoter regions for.}

\item{up.extend}{An integer specifying the number of base pairs upstream of
the TSS to include in the promoter region. Defaults to 2000.}

\item{dn.extend}{An integer specifying the number of base pairs downstream of
the TSS to include in the promoter region. Defaults to 0.}
}
\value{
A GRanges object containing the promoter regions for the specified genes or transcripts.
}
\description{
This method retrieves the promoter regions for specified genes or transcripts
in a GenomeGTF object. The promoter region is defined as the region upstream
of the transcription start site (TSS) of a gene or transcript.
}
\examples{
\dontrun{
# Retrieve promoter regions for gene "TP53" in GenomeGTF object "gtf":
promoters <- getPromoters(gtf, geneName = "TP53")

# Retrieve promoter regions for gene "BRCA1" and "BRCA2" in GenomeGTF object "gtf":
promoters <- getPromoters(gtf, geneName = c("BRCA1", "BRCA2"), up.extend = 5000)

# Retrieve promoter regions for transcript "ENST00000311936" in GenomeGTF object "gtf":
promoters <- getPromoters(gtf, transId = "ENST00000311936", up.extend = 1000, dn.extend = 1000)
}
}
