% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getPromoters}
\alias{getPromoters}
\title{getPromoters method for GenomeGTF objects}
\usage{
getPromoters(object, ...)
}
\arguments{
\item{object}{A GenomeGTF object}

\item{...}{Other arguments.}
}
\description{
getPromoters method for GenomeGTF objects
}
