% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRotatedPolygon}
\alias{getRotatedPolygon}
\title{getRotatedPolygon}
\usage{
getRotatedPolygon(
  data = NULL,
  rx = NULL,
  ry = NULL,
  value = NULL,
  theta = 45,
  workers = 1,
  window = 1
)
}
\arguments{
\item{data}{A data frame containing the original polygon coordinates.}

\item{rx}{The column name of the x-coordinates in the \code{data} parameter.}

\item{ry}{The column name of the y-coordinates in the \code{data} parameter.}

\item{value}{The column name of the values associated with each point in the polygon.}

\item{theta}{The angle in degrees to rotate the polygon.}

\item{workers}{The number of cores to use for parallel processing. Default is 1.}

\item{window}{The size of the window around the rotated polygon. Can be a numeric
value or the name of a column in the \code{data} parameter. Default is 1.}
}
\value{
A list containing the rotated coordinates and the polygon coordinates.
}
\description{
This function rotates a polygon based on a given angle and returns the rotated
coordinates along with a polygon of a given window size around the rotated polygon.
}
