% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getTransInfo,GenomeGTF-method}
\alias{getTransInfo,GenomeGTF-method}
\title{Get long name of transcripts in GenomeGTF object}
\usage{
\S4method{getTransInfo}{GenomeGTF}(
  object,
  geneName = NULL,
  geneId = NULL,
  transId = NULL,
  selecType = c("lcds", "lt"),
  topN = 1,
  sep = "|"
)
}
\arguments{
\item{object}{A GenomeGTF object.}

\item{geneName}{Character vector of gene names to filter by. Default is NULL.}

\item{geneId}{Character vector of gene IDs to filter by. Default is NULL.}

\item{transId}{Character vector of transcript IDs to filter by. Default is NULL.}

\item{selecType}{A character vector specifying the selection criteria.
Must be either "lt" for longest transcript or "lcds" for longest CDS. Default
is "lcds".}

\item{topN}{An integer indicating the number of longest transcripts to retain.
Default is 1. Set to 0 to retain all transcripts.}

\item{sep}{Separator character to use in the long name. Default is "|".}
}
\value{
A data frame.
}
\description{
This function extracts information about transcripts in a GenomeGTF object
and returns a data frame with their long names.
}
