% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{getTransInfo}
\alias{getTransInfo}
\title{getLongName method for GenomeGTF objects}
\usage{
getTransInfo(object, ...)
}
\arguments{
\item{object}{A GenomeGTF object}

\item{...}{Other arguments.}
}
\description{
getLongName method for GenomeGTF objects
}
