% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBigFile.R
\name{loadBed}
\alias{loadBed}
\title{Load BED files and create a data frame}
\usage{
loadBed(bed_file = NULL, file_name = NULL)
}
\arguments{
\item{bed_file}{A character vector of one or more file paths to BED files.}

\item{file_name}{A character vector of file name for re-assign.}
}
\value{
A data frame containing columns for chromosome, start position, end
position, sample name, and sample number.
}
\description{
This function takes one or more BED files and reads them into a data frame
containing columns for chromosome, start position, end position, sample name,
and sample number.
}
\examples{
\dontrun{bedfiles <- c("sample1.bed", "sample2.bed")
bed_df <- loadBed(bedfile = bedfiles)}

}
