% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFileFunctions.R
\name{loadBigWig}
\alias{loadBigWig}
\title{Load BigWig files and filter chromosomes}
\usage{
loadBigWig(bw_file = NULL, file_name = NULL, chrom = NULL, format = "bw")
}
\arguments{
\item{bw_file}{character vector of file path to BigWig files to be loaded}

\item{file_name}{character vector of sample name(s) corresponding to the BigWig
file(s). If NULL, the sample name will be extracted from the file path.}

\item{chrom}{character vector of chromosome name(s) to be included in the
output. If NULL, all chromosomes are included.}

\item{format}{the signal data format, "bw"(default), "wig" or "bedGraph".}
}
\value{
A data frame that contains the filtered data from the BigWig file(s)
and the sample name.
}
\description{
This function reads in one or multiple BigWig files and filters them by
chromosome. The output is a data frame that contains the data from the BigWig
file(s) and the sample name.
}
\examples{
\dontrun{
# Load all chromosomes from one BigWig file with a specified sample name
loadBigWig(bw_file = "/path/to/file.bw", file_name = "sample1")

# Load only specific chromosomes from multiple BigWig files with default sample names
loadBigWig(bw_file = c("/path/to/file1.bw", "/path/to/file2.bw"), chrom = c("chr1", "chr2"))
}

}
