% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{loadGenomeGTF}
\alias{loadGenomeGTF}
\title{loadGenomeGTF
Load genome annotation information from a GTF or GFF file and a genome fasta file.}
\usage{
loadGenomeGTF(
  gtfPath = NULL,
  genomePath = NULL,
  format = "gtf",
  filterProtein = FALSE
)
}
\arguments{
\item{gtfPath}{The path to the GTF or GFF annotation file.}

\item{genomePath}{The path to the genome fasta file.}

\item{format}{The annotation file format ("gtf" or "gff"), default is "gtf".}

\item{filterProtein}{Whether to filter protein-coding genes, default is FALSE.}
}
\value{
A GenomeGTF object containing genome annotation information.
}
\description{
loadGenomeGTF
Load genome annotation information from a GTF or GFF file and a genome fasta file.
}
\examples{
\dontrun{
# make object
mytest <- loadGenomeGTF(gtfPath = "hg38.ncbiRefSeq.gtf.gz",
                        genomePath = "hg38.fa.gz")
}
}
\author{
JunZhang
}
