% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBigFile.R
\name{loadloops}
\alias{loadloops}
\title{Load loop data from bed or bedpe format files}
\usage{
loadloops(loop_file = NULL, file_name = NULL, sep = "\\t")
}
\arguments{
\item{loop_file}{a character vector of file names to be loaded.}

\item{file_name}{a character vector of corresponding names to the file names
in \code{loop_file},or \code{NULL} to use the file names themselves as the
names of the output data frames.}

\item{sep}{the separator used in the input files.}
}
\value{
a data frame containing the loop data. The columns of the data frame
are as follows:
\itemize{
\item \code{seqnames}: chromosome names of the interacting regions;
\item \code{start}: mid-point of the first region;
\item \code{end}: mid-point of the second region;
\item \code{score}: a score calculated as the distance between the start and end positions
of the interacting regions divided by 100,000;
\item \code{fileName}: the name of the input file from which the loop data were extracted.
}
}
\description{
This function loads loop data from bed or bedpe format files and returns them
as a data frame.
}
\examples{
\dontrun{
loop_data <- loadloops(loop_file = c("loop1.bedpe", "loop2.bedpe"),
file_name = c("loop1", "loop2"),sep = "\t")
}

}
