% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawChromosome.R
\docType{data}
\name{mm10_obj}
\alias{mm10_obj}
\title{This is a test data for this package
test data describtion}
\format{
An object of class \code{list} of length 3.
}
\usage{
mm10_obj
}
\description{
This is a test data for this package
test data describtion
}
\author{
JunZhang
}
\keyword{datasets}
