% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{myShow,GenomeGTF-method}
\alias{myShow,GenomeGTF-method}
\title{Show method for GenomeGTF object}
\usage{
\S4method{myShow}{GenomeGTF}(object)
}
\arguments{
\item{object}{An object of class GenomeGTF.}
}
\value{
Nothing is returned; this function prints information to the console.
}
\description{
This function displays information about a GenomeGTF object, including the
status of its GTF and genome files.
}
\examples{
# Show information about the object
\dontrun{show(GenomeGTF)}

}
