% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{myShow}
\alias{myShow}
\title{myShow}
\usage{
myShow(object)
}
\arguments{
\item{object}{An object to myShow.}
}
\value{
None.
}
\description{
Define a generic method for showing objects
}
\details{
This function defines a generic method for showing objects, which can be
implemented for different classes of objects.
}
