% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepareKaryotype}
\alias{prepareKaryotype}
\title{prepareKaryotype}
\usage{
prepareKaryotype(chromSize = NULL, centromere = NULL)
}
\arguments{
\item{chromSize}{A data frame with columns "chr", "start", and "end" containing
the chromosome name, start position, and end position, respectively.}

\item{centromere}{A data frame with columns "chr", "ce_start", and "ce_end"
containing the chromosome name, centromere start position, and centromere end
position, respectively. If not supplied, the centromere start and end positions
will be set to 0.}
}
\value{
A data frame with columns "chr", "start", "end", "ce_start", and
"ce_end" containing the chromosome name, start position, end position,
centromere start position, and centromere end position, respectively.
}
\description{
This function creates a karyotype data frame for a given chromosome and
centromere size.
}
