% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyExtractSeq.R
\name{pyExtractSeq}
\alias{pyExtractSeq}
\title{Extract Sequence from Genome Based on Transcript ID}
\usage{
pyExtractSeq(
  gtf_file = NULL,
  genome_file = NULL,
  transcript_id = NULL,
  new_id = NULL,
  type = NULL,
  out_file = NULL,
  pythonPath = NULL
)
}
\arguments{
\item{gtf_file}{The path to the GTF file containing transcript information.}

\item{genome_file}{The path to the genome file where the sequence is to be extracted}

\item{transcript_id}{The transcript ID for the transcript to be extracted.}

\item{new_id}{The new ID to assign to the extracted sequence.}

\item{type}{The type of sequence to extract ('CDS', 'exon', or '3/5UTR').}

\item{out_file}{The output file to write the extracted sequence to.}

\item{pythonPath}{The path of python to be used for reticulate.}
}
\value{
A FASTA format string of the extracted sequence.
}
\description{
This function extracts the sequence of a transcript from the genome,
based on its specified in the GTF file.
}
\examples{
\dontrun{
pyExtractSeq(gtf_file = "transcripts.gtf", genome_file = "genome.fa",
transcript_id = "ENST00000456328", type = "cds", out_file = "extracted_seq.fasta")
}
}
