% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackVisProMax.R
\name{trackVisProMax}
\alias{trackVisProMax}
\title{Visualize genomic data using BioSeqUtils package}
\usage{
trackVisProMax(
  Input_gtf = NULL,
  Input_gene = NULL,
  upstream_extend = 0,
  downstream_extend = 0,
  Input_bw = NULL,
  Input_loop = NULL,
  Loop_curve_geom = "geom_arch2",
  Input_hic = NULL,
  Input_bed = NULL,
  Input_junction = NULL,
  query_region = list(query_chr = NULL, query_start = NULL, query_end = NULL),
  signal_layer_bw_params = list(),
  signal_layer_loop_params = list(),
  signal_layer_heatmap_params = list(),
  peaks_layer_params = list(),
  junc_layer_combined = FALSE,
  add_band_line = FALSE,
  band_width = 0.5,
  signal_layer_junction_params = list(),
  signal_layer_junction_label_params = list(),
  reverse_y_vars = NULL,
  gene_group_info = NULL,
  gene_group_info_order = NULL,
  gene_group_info2 = NULL,
  gene_group_info2_order = NULL,
  sample_group_info = NULL,
  sample_group_info_order = NULL,
  sample_group_info2 = NULL,
  sample_group_info2_order = NULL,
  gene_order = NULL,
  sample_order = NULL,
  draw_chromosome = FALSE,
  draw_chromosome_params = list(ideogram_obj = NULL),
  trans_topN = 2,
  collapse_trans = FALSE,
  exon_width = 0.5,
  peak_width = 0.5,
  add_gene_label_layer = FALSE,
  gene_label_shift_y = 0.3,
  gene_label_params = list(),
  arrow_rel_len_params_list = list(),
  trans_exon_arrow_params = list(),
  trans_exon_col_params = list(),
  gene_dist_mark_params = list(),
  gene_dist_mark_text_params = list(),
  signal_range_pos = c(0.85, 0.85),
  signal_range = NULL,
  panel_size_setting = list(),
  fixed_column_range = TRUE,
  signal_range_label_params = list(),
  higlight_region = NULL,
  higlight_col = NULL,
  higlight_col_alpha = 0.2,
  background_color_region = NULL,
  background_region_alpha = 0.2,
  by_layer_x = FALSE,
  by_layer_y = FALSE,
  column_strip_setting_list = list(),
  row_strip_setting_list = list(),
  column_strip_text_setting_list = list(),
  row_strip_text_setting_list = list(),
  add_gene_region_label = FALSE,
  base_size = 14,
  panel.spacing = c(0.2, 0),
  sample_fill_col = NULL,
  loops_col = NULL,
  heatmap_fill_col = NULL,
  peak_fill_col = NULL,
  trans_fill_col = NULL,
  remove_chrom_panel_border = FALSE,
  remove_all_panel_border = FALSE,
  xlimit_range = NULL,
  Intron_line_type = "line"
)
}
\arguments{
\item{Input_gtf}{A GTF file containing the gene annotation information
with data frame format. Must be specified if you want to add gene annotation tracks.}

\item{Input_gene}{A character vectors. Must be specified if you want to add}

\item{upstream_extend}{Extend bases upstream for gene transcription start site,
default 0, accepting one or more vectors.}

\item{downstream_extend}{Extend bases downstream for gene transcription end site,
default 0, accepting one or more vectors.
gene annotation tracks.}

\item{Input_bw}{A bigWig file from "loadBigWig" function containing the signal
information for signal tracks.}

\item{Input_loop}{A loop file from "loadloops" function containing the loop
information for loop tracks.}

\item{Loop_curve_geom}{The curve type for loop tracks, "geom_arch2"(default)
or "geom_arch".}

\item{Input_hic}{A Hi-C interaction file from "prepareHic" function containing
the Hi-C interaction information for heatmap tracks.}

\item{Input_bed}{A bed file from "loadBed" function containing the bed
information for bed tracks.}

\item{Input_junction}{A junction file from "loadJunction" function containing
the junction information for junction tracks.}

\item{query_region}{A named list specifying the region to be visualized.
It must contain three elements: query_chr, query_start, and query_end.}

\item{signal_layer_bw_params}{A list of parameters for configuring signal
tracks created by bigWig files. Passed by "ggplot2::geom_rect" function.}

\item{signal_layer_loop_params}{A list of parameters for configuring loop
tracks created by loop files. Passed by "ggbio::geom_arch" or "jjPlot::geom_arch2"
function.}

\item{signal_layer_heatmap_params}{A list of parameters for configuring
heatmap tracks created by Hi-C interaction files. Passed by
"ggplot2::geom_polygon" function.}

\item{peaks_layer_params}{A list of parameters for configuring bed tracks
created by bed files. Passed by "ggplot2::geom_rect" function.}

\item{junc_layer_combined}{A boolean indicating whether to combine splice
junctions from multiple samples into a single track. Passed by
"ggplot2::geom_rect" or "ggplot2::geom_polygon" function.}

\item{add_band_line}{A boolean indicating whether to add band-line for junction
curve.}

\item{band_width}{A numeric value specifying the width of band line.}

\item{signal_layer_junction_params}{A list of parameters for configuring
junction tracks created by junction files. Passed by
"ggbio::geom_arch" or "ggplot2::geom_polygon" function.}

\item{signal_layer_junction_label_params}{A list of parameters for configuring
labels for junction tracks. Passed by "ggplot2::geom_label" function.}

\item{reverse_y_vars}{Character vectors indicating whether to reverse
the y-axis for certain tracks.}

\item{gene_group_info}{A named list containing group information for genes.}

\item{gene_group_info_order}{Character vectors specifying the order of gene groups.}

\item{gene_group_info2}{A named list containing additional group information for genes.}

\item{gene_group_info2_order}{Character vectors specifying the order of
additional gene groups.}

\item{sample_group_info}{A named list containing group information for samples.}

\item{sample_group_info_order}{Character vectors specifying the order of sample groups.}

\item{sample_group_info2}{A named list containing additional group information for samples.}

\item{sample_group_info2_order}{Character vectors specifying the order of
additional sample groups.}

\item{gene_order}{Character vectors specifying the order of genes in the visualization.}

\item{sample_order}{Character vectors specifying the order of samples in the visualization.}

\item{draw_chromosome}{A boolean indicating whether to draw the chromosome
ideogram on the bottom of the plot.}

\item{draw_chromosome_params}{A list of parameters for configuring the chromosome ideogram.
Details see "drawChromosome" function.}

\item{trans_topN}{An integer specifying the number of top transcripts to be shown per gene.}

\item{collapse_trans}{A boolean indicating whether to collapse exons from the same transcript into a single block.}

\item{exon_width}{A numeric value specifying the width of each exon block.}

\item{peak_width}{A numeric value specifying the width of each peak block.}

\item{add_gene_label_layer}{A boolean indicating whether to add gene labels to the plot.}

\item{gene_label_shift_y}{A numeric value specifying the vertical shift of gene labels.}

\item{gene_label_params}{A list of parameters for configuring gene labels.
Passed by "ggplot2::geom_text" function.}

\item{arrow_rel_len_params_list}{A list of parameters for configuring arrow
relative length. Details see "createSegment" function.}

\item{trans_exon_arrow_params}{A list of parameters for configuring arrows
of a transcript. Passed by "ggplot2::geom_segment" function.}

\item{trans_exon_col_params}{A list of parameters for coloring exons. Supplying
with "fill" or "color" parameters.}

\item{gene_dist_mark_params}{A list of parameters for control the segment line
in the "trans" track. Passed by "ggplot2::geom_segment" function.}

\item{gene_dist_mark_text_params}{A list of parameters for controling the
textl label in the "trans" track. Passed by "ggplot2::geom_text" function.}

\item{signal_range_pos}{A two elements vector specifying the position of signal
range label in each track.}

\item{signal_range}{A named list specifying the range of signal values to be
displayed in signal tracks.}

\item{panel_size_setting}{A list of settings for each panel size. Details see
"ggh4x::force_panelsizes".}

\item{fixed_column_range}{A logical value indicating whether to use a fixed
column range.}

\item{signal_range_label_params}{A list of parameters for signal range label.
Passed by "ggplot2::geom_text" function.}

\item{higlight_region}{Null or a named list of named values representing the
highlighted region. Like "higlight_region <- list(Actb = list(start =
c(142904000),end = c(142904500)))".}

\item{higlight_col}{Null or a named list of values representing the color of
the highlighted region. Like "higlight_col <- list(Actb = c("grey50"))".}

\item{higlight_col_alpha}{A numeric value indicating the alpha level of the
highlighted region color.}

\item{background_color_region}{Null or a named list of named values representing
the background color of the region. Like "background_color_region <- list(Actb
= c("Input" = "grey90","IP-1"="yellow","IP-2"="purple"),
Myc = c("Input" = "red","IP-1"="blue","IP-2"="green"))"
"background_color_region <- list("chr5:65438603-65834288" = c("Input" = "grey90",
"IP-1"="yellow","IP-2"="purple"),"chr15:9263295-9785835" = c("Input" = "red",
"IP-1"="blue","IP-2"="green"))"}

\item{background_region_alpha}{A numeric value indicating the alpha level of
the background color of the region.}

\item{by_layer_x}{A logical value indicating whether to plot the strip by layer
in the x-axis.}

\item{by_layer_y}{A logical value indicating whether to plot the strip by layer
in the y-axis.}

\item{column_strip_setting_list}{A list of settings for column strips. Passed
with "ggh4x::elem_list_rect".}

\item{row_strip_setting_list}{A list of settings for row strips. Passed with
"ggh4x::elem_list_rect".}

\item{column_strip_text_setting_list}{A list of settings for column strip text.
Passed with "ggh4x::elem_list_text".}

\item{row_strip_text_setting_list}{A list of settings for row strip text.
Passed with "ggh4x::elem_list_text".}

\item{add_gene_region_label}{A logical value indicating whether to add genomic
region labels under the gene symbols.}

\item{base_size}{A numeric value indicating the base font size.}

\item{panel.spacing}{A numeric vector of values indicating the spacing between panels.}

\item{sample_fill_col}{Null or a numeric vector of values representing the fill
color of the samples.}

\item{loops_col}{Null or character vector of values representing the color of
the loops.}

\item{heatmap_fill_col}{Null or a character vector of values representing the
fill color of the heatmap.}

\item{peak_fill_col}{Null or a character vector of values representing the fill
color of the peaks.}

\item{trans_fill_col}{Null or a character vector of values representing the fill
color of the trans.}

\item{remove_chrom_panel_border}{A logical value indicating whether to remove
the panel border for the chromosome track.}

\item{remove_all_panel_border}{A logical value indicating whether to remove
all panel borders.}

\item{xlimit_range}{Null or a numeric vector of values representing the x-axis
limit range.}

\item{Intron_line_type}{Line type for intron regions, "line"(default) or "chevron".}
}
\value{
GGPLOT
}
\description{
Visualize genomic data using BioSeqUtils package
}
\references{
\url{https://junjunlab.github.io/BioSeqUtils-manual/}

\url{https://github.com/junjunlab/BioSeqUtils}
}
\author{
JunZhang

This function generates a multi-track visualization of different genomic data
types, including gene annotation, signal tracks, loops, Hi-C interactions, bed
files, and junction files. For more details and full documentation, please refer
to \href{https://junjunlab.github.io/BioSeqUtils-manual/}{BioSeqUtils-manual}.
}
