% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transCoordTransform}
\alias{transCoordTransform}
\title{Convert and transform coordinates of genomic features in a GTF file}
\usage{
transCoordTransform(gtf_file = NULL)
}
\arguments{
\item{gtf_file}{A data frame object with at least the columns "transcript_id",
"type", "strand", "start", "end", and "width".}
}
\value{
A data frame with the same columns as \code{gtf_file}, where the
coordinates of features are in transcript space.
}
\description{
This function takes a GTF file as input and converts the coordinates of
genomic features to transcript coordinates. It also reassigns the start and
end coordinates for exons, as well as for coding regions (CDS).
}
\examples{
\dontrun{
# Run function
transCoordTransform(gtf_file = example_gtf)
}

}
